Move coverart tests into test/coverart/
Relocate the top-level test_coverart*.py files into test/coverart/, matching
their subject module picard/coverart/.

The create_image() helper was defined in test_coverart_image.py and imported
by test_file, test_item and test_metadata. Move it to the shared
test/picardtestcase.py so tests no longer import from one another across
directories.
