+++ test/coverart/test_coverart_processing.py
