+++ test/coverart/test_coverart_queue.py
