+++ test/coverart/test_coverartprovider_caa.py
