+++ test/coverart/test_coverartprovider_local.py
