+++ test/test_file.py
@@ -30,8 +30,10 @@ from unittest.mock import (
     patch,
 )
 
-from test.picardtestcase import PicardTestCase
-from test.test_coverart_image import create_image
+from test.picardtestcase import (
+    PicardTestCase,
+    create_image,
+)
 
 from picard import config
 from picard.const.sys import (
